+++ src/gui/http.go
@@ -44,26 +44,30 @@ type Server struct {
 
 // Config configures Server
 type Config struct {
-	StaticDir        string
-	DisableCSRF      bool
-	DisableWalletAPI bool
-	ReadTimeout      time.Duration
-	WriteTimeout     time.Duration
-	IdleTimeout      time.Duration
+	StaticDir       string
+	DisableCSRF     bool
+	EnableWalletAPI bool
+	ReadTimeout     time.Duration
+	WriteTimeout    time.Duration
+	IdleTimeout     time.Duration
 }
 
 type muxConfig struct {
-	host             string
-	appLoc           string
-	disableWalletAPI bool
+	host            string
+	appLoc          string
+	enableWalletAPI bool
 }
 
 func create(host string, c Config, daemon *daemon.Daemon) (*Server, error) {
-	appLoc, err := file.DetermineResourcePath(c.StaticDir, resourceDir, devDir)
-	if err != nil {
-		return nil, err
+	var appLoc string
+	if c.EnableWalletAPI {
+		var err error
+		appLoc, err = file.DetermineResourcePath(c.StaticDir, resourceDir, devDir)
+		if err != nil {
+			return nil, err
+		}
+		logger.Infof("Web resources directory: %s", appLoc)
 	}
-	logger.Infof("Web resources directory: %s", appLoc)
 
 	csrfStore := &CSRFStore{
 		Enabled: !c.DisableCSRF,
@@ -83,9 +87,9 @@ func create(host string, c Config, daemon *daemon.Daemon) (*Server, error) {
 	}
 
 	mc := muxConfig{
-		host:             host,
-		appLoc:           appLoc,
-		disableWalletAPI: c.DisableWalletAPI,
+		host:            host,
+		appLoc:          appLoc,
+		enableWalletAPI: c.EnableWalletAPI,
 	}
 
 	srvMux := newServerMux(mc, daemon.Gateway, csrfStore)
@@ -184,7 +188,7 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 		mux.Handle(endpoint, handler)
 	}
 
-	if !c.disableWalletAPI {
+	if c.enableWalletAPI {
 		webHandler("/", newIndexHandler(c.appLoc))
 
 		fileInfos, _ := ioutil.ReadDir(c.appLoc)
@@ -266,10 +270,29 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	//     entropy: entropy bitsize.
 	webHandler("/wallet/newSeed", newWalletSeed(gateway))
 
+	// Gets seed of wallet of given id
+	// GET Arguments:
+	//     id: wallet id
+	//     password: wallet password
+	webHandler("/wallet/seed", walletSeedHandler(gateway))
+
 	// unload wallet
 	// POST Argument:
 	//         id: wallet id
 	webHandler("/wallet/unload", walletUnloadHandler(gateway))
+
+	// Encrypts wallet
+	// POST arguments:
+	//     id: wallet id
+	//     password: wallet password
+	// Returns an encrypted wallet json without sensitive data
+	webHandler("/wallet/encrypt", walletEncryptHandler(gateway))
+
+	// Decrypts wallet
+	// POST arguments:
+	//     id: wallet id
+	//     password: wallet password
+	webHandler("/wallet/decrypt", walletDecryptHandler(gateway))
 
 	// Blockchain interface
 
@@ -284,7 +307,6 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	webHandler("/last_blocks", getLastBlocks(gateway))
 
 	// Network stats interface
-
 	webHandler("/network/connection", connectionHandler(gateway))
 	webHandler("/network/connections", connectionsHandler(gateway))
 	webHandler("/network/defaultConnections", defaultConnectionsHandler(gateway))
@@ -297,6 +319,9 @@ func newServerMux(c muxConfig, gateway Gatewayer, csrfStore *CSRFStore) *http.Se
 	webHandler("/pendingTxs", getPendingTxs(gateway))
 	// get txn by txid
 	webHandler("/transaction", getTransactionByID(gateway))
+
+	// Health check handler
+	webHandler("/health", healthCheck(gateway))
 
 	// Returns transactions that match the filters.
 	// Method: GET