+++ src/visor/visor.go
@@ -75,6 +75,7 @@ func DropletPrecisionCheck(amount uint64) error {
 type BuildInfo struct {
 	Version string `json:"version"` // version number
 	Commit  string `json:"commit"`  // git commit id
+	Branch  string `json:"branch"`  // git branch name
 }
 
 // Config configuration parameters for the Visor
@@ -126,8 +127,10 @@ type Config struct {
 	WalletDirectory string
 	// build info, including version, build time etc.
 	BuildInfo BuildInfo
-	// disables wallet API
-	DisableWalletAPI bool
+	// enables wallet API
+	EnableWalletAPI bool
+	// enables seed API
+	EnableSeedAPI bool
 	// wallet crypto type
 	WalletCryptoType wallet.CryptoType
 }
@@ -238,9 +241,11 @@ type Visor struct {
 	Unconfirmed UnconfirmedTxnPooler
 	Blockchain  Blockchainer
 	Wallets     *wallet.Service
-	history     historyer
-	bcParser    *BlockchainParser
-	db          *bolt.DB
+	StartedAt   time.Time
+
+	history  historyer
+	bcParser *BlockchainParser
+	db       *bolt.DB
 }
 
 // NewVisor creates a Visor for managing the blockchain database
@@ -270,9 +275,10 @@ func NewVisor(c Config, db *bolt.DB) (*Visor, error) {
 	bc.BindListener(bp.FeedBlock)
 
 	wltServConfig := wallet.Config{
-		WalletDir:        c.WalletDirectory,
-		CryptoType:       c.WalletCryptoType,
-		DisableWalletAPI: c.DisableWalletAPI,
+		WalletDir:       c.WalletDirectory,
+		CryptoType:      c.WalletCryptoType,
+		EnableWalletAPI: c.EnableWalletAPI,
+		EnableSeedAPI:   c.EnableSeedAPI,
 	}
 
 	wltServ, err := wallet.NewService(wltServConfig)
@@ -288,6 +294,7 @@ func NewVisor(c Config, db *bolt.DB) (*Visor, error) {
 		history:     history,
 		bcParser:    bp,
 		Wallets:     wltServ,
+		StartedAt:   time.Now(),
 	}
 
 	return v, nil
@@ -546,7 +553,7 @@ func (vs *Visor) HeadBkSeq() uint64 {
 }
 
 // GetBlockchainMetadata returns descriptive Blockchain information
-func (vs *Visor) GetBlockchainMetadata() BlockchainMetadata {
+func (vs *Visor) GetBlockchainMetadata() (*BlockchainMetadata, error) {
 	return NewBlockchainMetadata(vs)
 }
 
@@ -1029,9 +1036,9 @@ func (vs Visor) GetAddrUxOuts(address cipher.Address) ([]*historydb.UxOut, error
 	return vs.history.GetAddrUxOuts(address)
 }
 
-// ScanAheadWalletAddresses scans ahead N addresses in a wallet, looking for a non-empty balance
-func (vs Visor) ScanAheadWalletAddresses(wltName string, password []byte, scanN uint64) (*wallet.Wallet, error) {
-	return vs.Wallets.ScanAheadWalletAddresses(wltName, password, scanN, vs)
+// CreateWallet creates wallet and scans ahead N addresses to look for a none-empty balance
+func (vs *Visor) CreateWallet(wltName string, opts wallet.Options) (*wallet.Wallet, error) {
+	return vs.Wallets.CreateWallet(wltName, opts, vs)
 }
 
 // GetBalanceOfAddrs returns balance pairs of given addreses